move separate_sets to utils
