+++ mahjong/yaku_types.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 from abc import ABC, abstractmethod
 
 from .player import Player
-from .utils import is_yaochuu
+from .utils import is_yaochuu, separate_sets
 from .components import Suit, Jihai, Tile, Naki
 from .naki_and_actions import check_tenpai
 
@@ -78,43 +78,6 @@ class YakuTypes(ABC):
                     fu += 16
                 else: fu += 8
 
-        def separate_sets(hand: DefaultDict[int, int], huro_count: int):
-
-            for possible_jantou in hand.keys():
-                if hand[possible_jantou] >= 2: # try using it as jantou
-                    remain_tiles = copy.deepcopy(hand)
-                    remain_tiles[possible_jantou] -= 2
-                    
-                    koutsu = []
-                    shuntsu = []
-                    sets_to_find = 4 - huro_count
-                    for tile_index in sorted(remain_tiles.keys()):
-                        if tile_index < Tile(Suit.MANZU.value, 1).index:  # only check Koutsu
-                            if remain_tiles[tile_index] == 3:
-                                sets_to_find -= 1
-                                koutsu.append(Tile.from_index(tile_index))
-                        else:  # numbered tiles
-                            if remain_tiles[tile_index] >= 3:  # check for Koutsu
-                                remain_tiles[tile_index] -= 3
-                                sets_to_find -= 1
-                                koutsu.append(Tile.from_index(tile_index))
-                            if remain_tiles[tile_index + 2] > 0:  # check for Shuntsu
-                                chii_n = min(remain_tiles[tile_index],
-                                            remain_tiles[tile_index + 1],
-                                            remain_tiles[tile_index + 2])
-                                if chii_n > 0:
-                                    remain_tiles[tile_index] -= chii_n
-                                    remain_tiles[tile_index + 1] -= chii_n
-                                    remain_tiles[tile_index + 2] -= chii_n
-                                    sets_to_find -= chii_n
-                                    for _ in range(chii_n):
-                                        shuntsu.append([
-                                            Tile.from_index(tile_index),
-                                            Tile.from_index(tile_index) + 1,
-                                            Tile.from_index(tile_index) + 2
-                                        ])
-                    if sets_to_find == 0:
-                        return koutsu, shuntsu, Tile.from_index(possible_jantou)
 
         def calc_wait_pattern_fu(ankous: List[Tile],
                                  shuntsus: List[Tile],
